+++ fitty_frontend_app/lib/home_page/BaseAppBar.dart
