+++ fitty_frontend_app/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'home_page/home_page.dart';
+import 'home_page/HomePage.dart';
 
 void main() {
   runApp(MyApp());
